+++ module/GeebyDeebyConsole/src/GeebyDeebyConsole/Command/PluginManager.php
@@ -4,7 +4,7 @@
  *
  * PHP version 7
  *
- * Copyright (C) Villanova University 2020.
+ * Copyright (C) Demian Katz 2020.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -46,44 +46,7 @@ class PluginManager extends \GeebyDeeby\ServiceManager\AbstractPluginManager
      * @var array
      */
     protected $aliases = [
-        'compile/theme' => Compile\ThemeCommand::class,
-        'generate/dynamicroute' => Generate\DynamicRouteCommand::class,
-        'generate/extendclass' => Generate\ExtendClassCommand::class,
-        'generate/extendservice' => Generate\ExtendServiceCommand::class,
-        'generate/nontabrecordaction' => Generate\NonTabRecordActionCommand::class,
-        'generate/plugin' => Generate\PluginCommand::class,
-        'generate/recordroute' => Generate\RecordRouteCommand::class,
-        'generate/staticroute' => Generate\StaticRouteCommand::class,
-        'generate/theme' => Generate\ThemeCommand::class,
-        'generate/thememixin' => Generate\ThemeMixinCommand::class,
-        'harvest/harvest_oai' => Harvest\HarvestOaiCommand::class,
-        'harvest/merge-marc' => Harvest\MergeMarcCommand::class,
-        'import/import-xsl' => Import\ImportXslCommand::class,
-        'import/webcrawl' => Import\WebCrawlCommand::class,
-        'install/install' => Install\InstallCommand::class,
-        'language/addusingtemplate' => Language\AddUsingTemplateCommand::class,
-        'language/copystring' => Language\CopyStringCommand::class,
-        'language/delete' => Language\DeleteCommand::class,
-        'language/normalize' => Language\NormalizeCommand::class,
-        'scheduledsearch/notify' => ScheduledSearch\NotifyCommand::class,
-        'util/cleanuprecordcache' => Util\CleanUpRecordCacheCommand::class,
-        'util/cleanup_record_cache' => Util\CleanUpRecordCacheCommand::class,
-        'util/commit' => Util\CommitCommand::class,
-        'util/createHierarchyTrees' => Util\CreateHierarchyTreesCommand::class,
-        'util/cssBuilder' => Util\CssBuilderCommand::class,
-        'util/dedupe' => Util\DedupeCommand::class,
-        'util/deletes' => Util\DeletesCommand::class,
-        'util/expire_auth_hashes' => Util\ExpireAuthHashesCommand::class,
-        'util/expire_external_sessions' => Util\ExpireExternalSessionsCommand::class,
-        'util/expire_searches' => Util\ExpireSearchesCommand::class,
-        'util/expire_sessions' => Util\ExpireSessionsCommand::class,
-        'util/index_reserves' => Util\IndexReservesCommand::class,
-        'util/lint_marc' => Util\LintMarcCommand::class,
-        'util/optimize' => Util\OptimizeCommand::class,
-        'util/scssBuilder' => Util\ScssBuilderCommand::class,
-        'util/sitemap' => Util\SitemapCommand::class,
-        'util/suppressed' => Util\SuppressedCommand::class,
-        'util/switch_db_hash' => Util\SwitchDbHashCommand::class,
+        'check/links' => Check\LinksCommand::class
     ];
 
     /**
@@ -92,62 +55,7 @@ class PluginManager extends \GeebyDeeby\ServiceManager\AbstractPluginManager
      * @var array
      */
     protected $factories = [
-        Compile\ThemeCommand::class => Compile\ThemeCommandFactory::class,
-        Generate\DynamicRouteCommand::class =>
-            Generate\AbstractRouteCommandFactory::class,
-        Generate\ExtendClassCommand::class =>
-            Generate\AbstractContainerAwareCommandFactory::class,
-        Generate\ExtendServiceCommand::class =>
-            Generate\AbstractCommandFactory::class,
-        Generate\NonTabRecordActionCommand::class =>
-            Generate\NonTabRecordActionCommandFactory::class,
-        Generate\PluginCommand::class =>
-            Generate\AbstractContainerAwareCommandFactory::class,
-        Generate\RecordRouteCommand::class =>
-            Generate\AbstractRouteCommandFactory::class,
-        Generate\StaticRouteCommand::class =>
-            Generate\AbstractRouteCommandFactory::class,
-        Generate\ThemeCommand::class =>
-            Generate\ThemeCommandFactory::class,
-        Generate\ThemeMixinCommand::class =>
-            Generate\ThemeMixinCommandFactory::class,
-        Harvest\MergeMarcCommand::class => InvokableFactory::class,
-        Harvest\HarvestOaiCommand::class => Harvest\HarvestOaiCommandFactory::class,
-        Import\ImportXslCommand::class => Import\ImportXslCommandFactory::class,
-        Import\WebCrawlCommand::class => Import\WebCrawlCommandFactory::class,
-        Install\InstallCommand::class => InvokableFactory::class,
-        Language\AddUsingTemplateCommand::class =>
-            Language\AbstractCommandFactory::class,
-        Language\CopyStringCommand::class => Language\AbstractCommandFactory::class,
-        Language\DeleteCommand::class => Language\AbstractCommandFactory::class,
-        Language\NormalizeCommand::class => Language\AbstractCommandFactory::class,
-        ScheduledSearch\NotifyCommand::class =>
-            ScheduledSearch\NotifyCommandFactory::class,
-        Util\CleanUpRecordCacheCommand::class =>
-            Util\CleanUpRecordCacheCommandFactory::class,
-        Util\CommitCommand::class => Util\AbstractSolrCommandFactory::class,
-        Util\CreateHierarchyTreesCommand::class =>
-        Util\CreateHierarchyTreesCommandFactory::class,
-        Util\CssBuilderCommand::class => Util\CssBuilderCommandFactory::class,
-        Util\DedupeCommand::class => InvokableFactory::class,
-        Util\DeletesCommand::class => Util\AbstractSolrCommandFactory::class,
-        Util\ExpireAuthHashesCommand::class =>
-            Util\ExpireAuthHashesCommandFactory::class,
-        Util\ExpireExternalSessionsCommand::class =>
-            Util\ExpireExternalSessionsCommandFactory::class,
-        Util\ExpireSearchesCommand::class =>
-            Util\ExpireSearchesCommandFactory::class,
-        Util\ExpireSessionsCommand::class =>
-            Util\ExpireSessionsCommandFactory::class,
-        Util\IndexReservesCommand::class =>
-            Util\AbstractSolrAndIlsCommandFactory::class,
-        Util\LintMarcCommand::class => InvokableFactory::class,
-        Util\OptimizeCommand::class => Util\AbstractSolrCommandFactory::class,
-        Util\ScssBuilderCommand::class => Util\ScssBuilderCommandFactory::class,
-        Util\SitemapCommand::class => Util\SitemapCommandFactory::class,
-        Util\SuppressedCommand::class =>
-            Util\AbstractSolrAndIlsCommandFactory::class,
-        Util\SwitchDbHashCommand::class => Util\SwitchDbHashCommandFactory::class,
+        Check\LinksCommand::class => Check\LinksCommandFactory::class,
     ];
 
     /**