+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- Command line interface for running tools (accessible through cli.php)
+- Command line utility for checking status of full text links.
 
 ### Changed